src/main/java/com/moviereservation/utils/strategy/FixDiscountPolicy.java
package com.moviereservation.utils.strategy;

import com.moviereservation.domain.member.Grade;
import com.moviereservation.domain.member.Member;
import org.springframework.stereotype.Component;

@Component
public class FixDiscountPolicy implements DiscountPolicy{

    private int discountAmount = 1000;

    @Override
    public int discount(Member member, int price) {
        if (member.getGrade() == Grade.VIP) {
            return discountAmount;
        } else {
            return 0;
        }
    }
}
